move context processors to another file and fix test
