+++ main.py
@@ -3,6 +3,7 @@ from app import app
 from app import app_config
 from config import constants
 from util import patches
+from util.context import create_contexts
 
 from flask_compress import Compress
 
@@ -21,6 +22,9 @@ assert campaign_views
 Compress(app)
 
 app_config.init_prod_app(app)
+
+# Template context processors
+create_contexts(app)
 
 migrate = Migrate(app, db)
 
@@ -41,4 +45,9 @@ if __name__ == '__main__':
 
     app.debug = constants.DEBUG
     port = int(os.environ.get("PORT", 5000))
-    app.run(host='0.0.0.0', port=port, threaded=True, extra_files=extra_files)
+    app.run(
+        host='0.0.0.0',
+        port=port,
+        threaded=True,
+        extra_files=extra_files
+    )
